Remove .float() when sending to cuda

+++ summarizer/models/baseline.py
@@ -68,7 +68,7 @@ class LogisticRegressionModel(Model):
                 target /= target.max()
 
                 if self.hps.use_cuda:
-                    seq, target = seq.float().cuda(), target.float().cuda()
+                    seq, target = seq.cuda(), target.cuda()
 
                 y = self.model(seq)
 
@@ -105,7 +105,7 @@ class LogisticRegressionModel(Model):
                 seq = torch.from_numpy(seq).unsqueeze(0)
 
                 if self.hps.use_cuda:
-                    seq = seq.float().cuda()
+                    seq = seq.cuda()
 
                 y = self.model(seq)
                 summary[key] = y[0].detach().cpu().numpy()
